feature/swagger

+++ ftp/src/FTPClient.ts
@@ -3,6 +3,13 @@ import {NextFunction, Request, Response} from 'express';
 import path from 'path';
 import * as fs from 'fs';
 
+interface IErrorResponseOptions {
+    res: Response,
+    statusCode: number,
+    errorMessage?: string,
+    invalidQueryParams?: boolean
+}
+
 export default class FTPClient {
     private client: Client;
 
@@ -11,6 +18,14 @@ export default class FTPClient {
         this.client.connect(clientOptions);
     }
 
+    public errorRespnse = (opts: IErrorResponseOptions) => {
+        const {res, statusCode, errorMessage, invalidQueryParams} = opts;
+        res.status(statusCode).json({
+            isSuccess: false,
+            error: invalidQueryParams ? 'please provide the required query params' : errorMessage
+        });
+    };
+
     /**
      * List the content from the specified path on FTP server
      * @param req
@@ -18,14 +33,18 @@ export default class FTPClient {
      * @param next
      */
     public list(req: Request, res: Response, next: NextFunction) {
-        this.client.list(req.params.path, (err, list) => {
-            try {
-                if (err) throw err;
-                return res.json({isSuccess: true, data: list});
-            } catch (error) {
-                res.json({isSuccess: false, error: error.message});
-            }
-        });
+        if (req.query.path) {
+            this.client.list((req.query.path as string), (err, list) => {
+                try {
+                    if (err) throw err;
+                    return res.json({isSuccess: true, data: list});
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
     }
 
     /**
@@ -34,18 +53,22 @@ export default class FTPClient {
      * @param res
      * @param next
      */
-    public download(req: Request, res: Response, next: NextFunction) {
+    public download = (req: Request, res: Response, next: NextFunction) => {
         console.log(`downloading from FTP server path: ${req.params.path}`);
-        this.client.get(req.params.path, (err, stream) => {
-            try {
-                if (err) throw err;
-                res.setHeader('content-type', 'some/type');
-                stream.pipe(res);
-            } catch (error) {
-                res.send({isSuccess: false, error: error.message});
-            }
-        });
-    }
+        if (req.query.path) {
+            this.client.get((req.query.path as string), (err, stream) => {
+                try {
+                    if (err) throw err;
+                    res.setHeader('content-type', 'some/type');
+                    stream.pipe(res);
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
+    };
 
     /**
      * Uploads the file from previously saved multer temporary location directly to a specified FTP server path
@@ -56,19 +79,23 @@ export default class FTPClient {
     public upload(req: Request, res: Response, next: NextFunction) {
         // path to temporary multer api storage
         const tempPath = path.resolve(__dirname, `../temp/${req.file?.originalname}`);
-        this.client.put(fs.createReadStream(tempPath), `${req.params.path}/${req.file!.originalname}`, err => {
-            try {
-                if (err) throw err;
-                // delete temporary file as it is already saved on our FTP server
-                fs.unlinkSync(tempPath);
-                res.send({
-                    isSuccess: true,
-                    message: `file: ${req.file?.originalname} successfully uploaded to FTP server!`
-                });
-            } catch (error) {
-                res.send({isSuccess: false, error: error.message});
-            }
-        });
+        if (req.query.path) {
+            this.client.put(fs.createReadStream(tempPath), `${req.query.path}/${req.file!.originalname}`, err => {
+                try {
+                    if (err) throw err;
+                    // delete temporary file as it is already saved on our FTP server
+                    fs.unlinkSync(tempPath);
+                    res.send({
+                        isSuccess: true,
+                        message: `file: ${req.file?.originalname} successfully uploaded to FTP server!`
+                    });
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
     }
 
     /**
@@ -78,17 +105,21 @@ export default class FTPClient {
      * @param next
      */
     public delete(req: Request, res: Response, next: NextFunction) {
-        this.client.delete(req.params.path, (err) => {
-            try {
-                if (err) throw err;
-                res.send({
-                    isSuccess: true,
-                    message: `file: ${req.params.path} was successfully deleted!`
-                });
-            } catch (error) {
-                res.send({isSuccess: false, error: error.message});
-            }
-        });
+        if (req.query.path) {
+            this.client.delete((req.query.path as string), (err) => {
+                try {
+                    if (err) throw err;
+                    res.send({
+                        isSuccess: true,
+                        message: `file: ${req.params.path} was successfully deleted!`
+                    });
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
     }
 
     /**
@@ -98,18 +129,21 @@ export default class FTPClient {
      * @param next
      */
     public mkdir(req: Request, res: Response, next: NextFunction) {
-        this.client.mkdir(req.params.path, true, (err) => {
-            try {
-                if (err) throw err;
-                res.send({
-                    isSuccess: true,
-                    message: `directory ${req.params.path} was successfully created!`
-                });
-            } catch (error) {
-                res.send({isSuccess: false, error: error.message});
-            }
-        });
-
+        if (req.query.path) {
+            this.client.mkdir((req.query.path as string), true, (err) => {
+                try {
+                    if (err) throw err;
+                    res.send({
+                        isSuccess: true,
+                        message: `directory ${req.params.path} was successfully created!`
+                    });
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
     }
 
     /**
@@ -119,17 +153,21 @@ export default class FTPClient {
      * @param next
      */
     public rmdir(req: Request, res: Response, next: NextFunction) {
-        this.client.rmdir(req.params.path, false, (err) => {
-            try {
-                if (err) throw err;
-                res.send({
-                    isSuccess: true,
-                    message: `directory ${req.params.path} was successfully deleted!`
-                });
-            } catch (error) {
-                res.send({isSuccess: false, error: error.message});
-            }
-        });
+        if (req.query.path) {
+            this.client.rmdir((req.query.path as string), false, (err) => {
+                try {
+                    if (err) throw err;
+                    res.send({
+                        isSuccess: true,
+                        message: `directory ${req.params.path} was successfully deleted!`
+                    });
+                } catch (error) {
+                    this.errorRespnse({res, statusCode: 500, errorMessage: error.message});
+                }
+            });
+        } else {
+            this.errorRespnse({res, statusCode: 400, invalidQueryParams: true});
+        }
     }
 
 }
